Wrap lists in models add station checksum method
Wraps lists from the API in list model classes to allow preservation of
additional attributes such as the checksum in the case of stations. This
should also protect us in the future if new attributes are added to
items that are currently lists.

All list models are actual list subclasses so they should behave
correctly for anyone expecting real lists.

fixes #4