Support new-style audio URLs
Pandora now returns two different responses to the API depending on
which API key the client is using and the tuner endpoint. Instead of a
quality map only a single audio URL is returned which is of AAC SBR
format. This change accommodates that and returns the proper bitrate and
format based on empirical testing.

see: #52