Split player backends, add VLC strategy
Extract core player logic and mpg123-bound logic into parent-child
classes so that other player backend strategies can be added. Create a
headless VLC strategy that uses VLC if it's available. Update the pydora
player to prefer VLC if it's available on the system because it supports
a much more broad set of codecs and Pandora is now preferring AAC
formatted files.