Check if termios is present or ignore
termios is POSIX specific which doesn't work for Windows users using the
CLI. Check for the presence of termios and disable the set_echo function
if it's not present.

The only reason echo is disabled during polling is to prevent control
characters entered by a user from displaying in the playlist so
it's more cosmetic than anything.