Remove PyCrypto and add Cryptography
PyCrypto is no longer supported nor recommended. Cryptography is the
replacement package which is actively maintained by the Python core
committers. Also Cryptography is bundled as binary wheels for all the
platforms we support which removes the install-tiem compiler
requirement, making it easier to distribute Pydora. This commit replaces
all usages of PyCrypto with Cryptography but is completely API
compatible with the previous version.