Refactor audio URL extraction
This generalizes the audio URL extraction logic into a synthesized field
that can be extracted from the PlaylistModel class. It also removes a
few transmogrifiers that are no longer needed in the general case.

Technically this breaks a publicly exposed API within playlists models
but it was always considered an implementation detail so nobody should
be relying on it.